refactor: Take API url as argument, don't generate
Because Populi's API urls may change (especially between staging and
production environments), it is easier to not try and be smart about
building a URL and instead allow the consumer to provide one as
configuration.

Replaces the method signature for Populi.connect() to use url: instead
of school: as a named argument.

Also updates docs re: authorization in README